Move Egg Dockerfile into egg module
