+++ leash.gemspec
@@ -31,7 +31,7 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency "bundler", "~> 1.15"
   spec.add_development_dependency "rake", "~> 12.0"
-  spec.add_development_dependency "rspec", '~> 3.6', '>= 3.6.0'
+  spec.add_development_dependency "rspec", "~> 3.6", ">= 3.6.0"
   spec.add_development_dependency "rspec_junit_formatter", "~> 0.2.3"
   spec.add_development_dependency "pronto", "~> 0.9.3"
   spec.add_development_dependency "pronto-rubocop", "~> 0.9.0"
