+++ lib/egg.rb
@@ -14,3 +14,4 @@ require_relative "egg/cli"
 require_relative "egg/templates"
 require_relative "egg/configuration"
 require_relative "egg/docker_compose"
+require_relative "egg/dockerfile"
