+++ spec/lib/dockerfile/ruby_spec.rb
@@ -1,28 +1,28 @@
-require "dockerfile"
+require "egg/dockerfile"
 
-RSpec.describe Dockerfile::Ruby do
+RSpec.describe Egg::Dockerfile::Ruby do
   it "Is Available from #use" do
-    expect(Dockerfile.use("Ruby")).to be_an_instance_of(Dockerfile::Ruby)
+    expect(Egg::Dockerfile.use("Ruby")).to be_an_instance_of(Egg::Dockerfile::Ruby)
   end
 
   it "Can render from a template" do
-    df = Dockerfile.use "Ruby"
+    df = Egg::Dockerfile.use "Ruby"
     df.command = ["bin/rails", "server", "-p 3000"]
     df.ruby_version = "2.4.0"
     expect(df.render).to match(/FROM ruby/)
   end
 
   it "Requires setting the ruby_version" do
-    df = Dockerfile.use "Ruby"
+    df = Egg::Dockerfile.use "Ruby"
     df.command = ["bin/rails", "server", "-p 3000"]
-    expect { df.render }.to raise_error(Dockerfile::MissingPropertyError)
+    expect { df.render }.to raise_error(Egg::Dockerfile::MissingPropertyError)
       .with_message(/ruby_version/)
   end
 
   it "Requires setting the command" do
-    df = Dockerfile.use "Ruby"
+    df = Egg::Dockerfile.use "Ruby"
     df.ruby_version = "2.4.0"
-    expect { df.render }.to raise_error(Dockerfile::MissingPropertyError)
+    expect { df.render }.to raise_error(Egg::Dockerfile::MissingPropertyError)
       .with_message(/command/)
   end
 end
