+++ spec/lib/dockerfile_spec.rb
@@ -1,21 +1,21 @@
 require "spec_helper"
-require "dockerfile"
+require "egg/dockerfile"
 
-RSpec.describe Dockerfile do
+RSpec.describe Egg::Dockerfile do
   it "can be configured with a #use" do
-    df = Dockerfile.use("NodeJS")
-    expect(df).to be_an_instance_of(Dockerfile::NodeJS)
+    df = Egg::Dockerfile.use("NodeJS")
+    expect(df).to be_an_instance_of(Egg::Dockerfile::NodeJS)
   end
 
   it "Doesn't allow #use for templates that dont exist" do
     expect do
-      Dockerfile.use("FooLang")
-    end.to raise_error(Dockerfile::NoDockerfileError)
+      Egg::Dockerfile.use("FooLang")
+    end.to raise_error(Egg::Dockerfile::NoDockerfileError)
   end
 
   describe "#run" do
     it "Appends another RUN action" do
-      df = Dockerfile.use("Ruby")
+      df = Egg::Dockerfile.use("Ruby")
       df.command = "test"
       df.ruby_version = "2.3.3"
       df.run "echo awesomesauce"
@@ -23,7 +23,7 @@ RSpec.describe Dockerfile do
     end
 
     it "Allows you to run some commands before a specific command" do
-      df = Dockerfile.use("Ruby")
+      df = Egg::Dockerfile.use("Ruby")
       df.command = "test"
       df.ruby_version = "2.3.3"
       df.run "echo before bundling", before: [:run, "bundle install"]
